requirements.txt
streamlit
pyzbar
Pillow